+++ src/customers/customers.routes.ts
@@ -4,7 +4,8 @@ import multer from 'multer';
 import { CustomerController } from './controller';
 import { authenticateJWT } from '../middlewares/auth.middleware';
 import validateDto from '../common/utils/validate.dto';
-import { CreateCustomerDto, UpdateCustomerDto } from './dto/customer.dto';
+import { CreateCustomerDto } from './dto/create-customer.dto';
+import { UpdateCustomerDto } from './dto/update-customer.dto';
 
 const storage = multer.memoryStorage();
 const upload = multer({
